portal-service-test Move ContactLocalServiceTest of com.liferay.portal.service package to contacts-test module
